refactor!: rename statusline to cosmosline
