+++ lua/feline-cosmos/cosmosline/init.lua
@@ -51,8 +51,8 @@ local function patch(t1, t2)
 end
 
 M.setup = function(customization, themes)
-    local dark = themes and themes.dark or require('feline-cosmos.statusline.themes.dark')
-    local light = themes and themes.light or require('feline-cosmos.statusline.themes.light')
+    local dark = themes and themes.dark or require('feline-cosmos.cosmosline.themes.dark')
+    local light = themes and themes.light or require('feline-cosmos.cosmosline.themes.light')
 
     local statusline = {
         active = { left = active_left, middle = active_middle, right = active_right },
@@ -63,7 +63,7 @@ M.setup = function(customization, themes)
 
     local config = require('feline-cosmos').setup({
         theme = dark,
-        vi_mode_colors = require('feline-cosmos.statusline.themes.vi_mode_colors'),
+        vi_mode_colors = require('feline-cosmos.cosmosline.themes.vi_mode_colors'),
         components = {
             active = {
                 statusline.active.left,
@@ -85,7 +85,7 @@ M.setup = function(customization, themes)
 
     vim.cmd([[augroup cosmos_themes
         autocmd!
-        autocmd ColorScheme * lua require('feline-cosmos.statusline').select_theme()
+        autocmd ColorScheme * lua require('feline-cosmos.cosmosline').select_theme()
     augroup END]])
 
     return config
