+++ lua/feline-cosmos/cosmosline/themes/light.lua
