+++ lua/feline-cosmos/cosmosline/themes/vi_mode_colors.lua
